added dtype float 32

+++ Env/NormalizeReward.py
@@ -18,8 +18,8 @@ class RunningMeanStd:
         self.count = epsilon
 
     def update(self, x):
-        batch_mean = np.mean(x)
-        batch_var = np.var(x)
+        batch_mean = np.mean(x, dtype=np.float32)
+        batch_var = np.var(x, dtype=np.float32)
         batch_count = x.shape[0]
         self.update_from_moments(batch_mean, batch_var, batch_count)
 
@@ -39,7 +39,7 @@ def update_mean_var_count_from_moments(
     new_mean = mean + delta * batch_count / tot_count
     m_a = var * count
     m_b = batch_var * batch_count
-    M2 = m_a + m_b + np.square(delta) * count * batch_count / tot_count
+    M2 = m_a + m_b + np.square(delta, dtype=np.float32) * count * batch_count / tot_count
     new_var = M2 / tot_count
     new_count = tot_count
 
@@ -59,14 +59,14 @@ class NormalizeReward(gym.core.Wrapper):
         self.num_envs = getattr(env, "num_envs", 1)
         self.is_vector_env = getattr(env, "is_vector_env", False)
         self.return_rms = RunningMeanStd(mean, var)
-        self.returns = np.zeros(self.num_envs)
+        self.returns = np.zeros(self.num_envs, dtype=np.float32)
         self.gamma = gamma
         self.epsilon = epsilon
 
     def step(self, action):
         obs, rews, dones, infos = self.env.step(action)
         if not self.is_vector_env:
-            rews = np.array([rews])
+            rews = np.array([rews], dtype=np.float32)
         self.returns = self.returns * self.gamma + rews
         rews = self.normalize(rews)
         self.returns[dones] = 0.0
@@ -76,7 +76,7 @@ class NormalizeReward(gym.core.Wrapper):
 
     def normalize(self, rews):
         self.return_rms.update(self.returns)
-        return rews / np.sqrt(self.return_rms.var + self.epsilon)
+        return rews / np.sqrt(self.return_rms.var + self.epsilon, dtype=np.float32)
 
     def get_monitor_data(self):
         return self.env.get_monitor_data()
